requirements.txt
boto3==1.16.29
botocore==1.19.29
jmespath==0.10.0
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.15.0
urllib3==1.26.2
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
boto3==1.16.29
botocore==1.19.29
certifi==2020.12.5
chardet==3.0.4
geoip2~=2.4.2
idna==2.10
jmespath==0.10.0
maxminddb==2.0.3
multidict==5.1.0
python-dateutil==2.8.1
requests==2.25.0
s3transfer==0.3.3
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.2
yarl==1.6.3
PyYAML~=3.11